+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 0.0.7 (Not Released Yet)
 
 * On retirer l'ancien filtre «problèmes» dans la recherche de transaction.
+* Remplacement du trigger Civirule new_campagnodon_transaction par un trigger de type cron «X jours après».
 
 ## 0.0.6